created Topbar as an interface and BasicTopbar
